feature: triage root add command
